threads-bot/main.py
"""
Threads 自动化机器人 - 主程序
用途：24/7 自动经营 Threads 账号
"""

import os
import time
import random
from datetime import datetime
from dotenv import load_dotenv

# 加载配置
load_dotenv('config/.env')

class ThreadsBot:
    """Threads 自动化机器人"""
    
    def __init__(self):
        self.username = os.getenv('THREADS_USERNAME')
        self.password = os.getenv('THREADS_PASSWORD')
        self.auto_approve = os.getenv('AUTO_APPROVE', 'false').lower() == 'true'
        
        print("🤖 Threads Bot 初始化中...")
        print("🤖 Threads Bot initializing...")
        
    def learn_style(self):
        """学习用户发文风格"""
        print("\n📚 开始学习风格...")
        print("📚 Learning your style...")
        
        # TODO: 实现风格学习
        # 1. 抓取 IG 和 Threads 历史贴文
        # 2. AI 分析风格
        # 3. 生成风格模型
        
        pass
    
    def generate_post(self):
        """生成新贴文"""
        print("\n✍️ 生成新贴文...")
        print("✍️ Generating new post...")
        
        # TODO: 实现内容生成
        # 1. 根据风格模型生成
        # 2. 生成 3 个候选
        # 3. 选择最佳
        
        pass
    
    def auto_like(self):
        """自动点赞"""
        print("\n❤️ 自动点赞中...")
        print("❤️ Auto-liking posts...")
        
        # TODO: 实现自动点赞
        # 1. 获取热门贴文
        # 2. 随机选择
        # 3. 点赞并记录
        
        pass
    
    def auto_reply(self):
        """自动回复"""
        print("\n💬 自动回复中...")
        print("💬 Auto-replying to posts...")
        
        # TODO: 实现自动回复
        # 1. 找出热门贴文
        # 2. 生成回复
        # 3. 发送回复
        
        pass
    
    def run(self):
        """主循环"""
        print("\n🚀 Threads Bot 启动！")
        print("🚀 Threads Bot started!")
        
        while True:
            try:
                current_hour = datetime.now().hour
                
                # 每天 11:00 发文
                if current_hour == 11:
                    self.generate_post()
                
                # 随机时间点赞和回复
                if random.random() < 0.1:  # 10% 概率
                    self.auto_like()
                
                if random.random() < 0.05:  # 5% 概率
                    self.auto_reply()
                
                # 休眠随机时间
                sleep_time = random.randint(300, 3600)
                print(f"\n😴 休眠 {sleep_time} 秒...")
                print(f"😴 Sleeping for {sleep_time} seconds...")
                time.sleep(sleep_time)
                
            except KeyboardInterrupt:
                print("\n\n👋 Threads Bot 已停止")
                print("👋 Threads Bot stopped")
                break
            except Exception as e:
                print(f"\n❌ 错误：{e}")
                print(f"❌ Error: {e}")
                time.sleep(60)

if __name__ == "__main__":
    bot = ThreadsBot()
    bot.run()
